requirements.txt
fastapi
google-adk
uvicorn
psycopg2-binary
requests